# Grp2BackendAN/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "thinkbridge.Grp2BackendAN.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/ToDoChecklist/GetAllToDoChecklist": {
      "post": {
        "tags": [
          "ToDoChecklist"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoChecklistReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoChecklistReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoChecklistReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllToDoChecklistReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoChecklist/GetToDoChecklistById/{ToDoChecklistId}/{withDetails}": {
      "get": {
        "tags": [
          "ToDoChecklist"
        ],
        "parameters": [
          {
            "name": "ToDoChecklistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "withDetails",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDetailDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDetailDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDetailDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoChecklist/AddToDoChecklist": {
      "post": {
        "tags": [
          "ToDoChecklist"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoChecklistReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoChecklistReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoChecklistReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddToDoChecklistReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoChecklist/UpdateToDoChecklist": {
      "post": {
        "tags": [
          "ToDoChecklist"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoChecklistReqDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoChecklistReqDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoChecklistReqDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToDoChecklistReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoChecklistResDtoSingleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ToDoChecklist/DeleteToDoChecklist/{ToDoChecklistId}": {
      "delete": {
        "tags": [
          "ToDoChecklist"
        ],
        "parameters": [
          {
            "name": "ToDoChecklistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddToDoChecklistReqDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "BooleanFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BooleanFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BooleanFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "boolean"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "ComparisonOperator": {
        "enum": [
          "Equals",
          "NotEquals",
          "GreaterThan",
          "LessThan",
          "GreaterThanOrEqual",
          "LessThanOrEqual",
          "Contains"
        ],
        "type": "string"
      },
      "DateTimeFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateTimeFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateTimeFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "date-time"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "DateTimeNullableFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateTimeNullableFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateTimeNullableFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
...
      "GetAllToDoChecklistReqDto": {
        "type": "object",
        "properties": {
          "toDoChecklistId": {
            "$ref": "#/components/schemas/Int32FilterExpression"
          },
          "title": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "description": {
            "$ref": "#/components/schemas/StringFilterExpression"
          },
          "isComplete": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "createdDate": {
            "$ref": "#/components/schemas/DateTimeFilterExpression"
          },
          "updatedDate": {
            "$ref": "#/components/schemas/DateTimeNullableFilterExpression"
          },
          "toDoTaskId": {
            "$ref": "#/components/schemas/Int32FilterExpression"
          },
          "isDeleted": {
            "$ref": "#/components/schemas/BooleanFilterExpression"
          },
          "options": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "additionalProperties": false
      },
...
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "Int32FilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Int32FilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Int32FilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
      "LogicalOperator": {
        "enum": [
          "And",
          "Or"
        ],
        "type": "string"
      },
      "PageDetails": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sortBy": {
            "type": "string",
            "nullable": true
          },
          "sortDescending": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ResponseMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringFilterExpression": {
        "type": "object",
        "properties": {
          "combineWith": {
            "$ref": "#/components/schemas/LogicalOperator"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringFilterOption"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringFilterOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "comparisonOperator": {
            "$ref": "#/components/schemas/ComparisonOperator"
          }
        },
        "additionalProperties": false
      },
...
      "ToDoChecklistResDetailDto": {
        "type": "object",
        "properties": {
          "toDoChecklistToDoTaskIdfk": {
            "$ref": "#/components/schemas/ToDoMasterResDto"
          },
          "toDoChecklistId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ToDoChecklistResDetailDtoSingleResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToDoChecklistResDetailDto"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoChecklistResDto": {
        "type": "object",
        "properties": {
          "toDoChecklistId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ToDoChecklistResDtoListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToDoChecklistResDto"
            },
            "nullable": true
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageDetails"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ToDoChecklistResDtoSingleResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToDoChecklistResDto"
          },
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ToDoMasterResDto": {
        "type": "object",
        "properties": {
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "repeatFrequency": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "isAssignedToTeam": {
            "type": "boolean"
          },
          "hasChecklist": {
            "type": "boolean"
          },
          "hasReminder": {
            "type": "boolean"
          },
          "percentageCompleted": {
            "type": "number",
            "format": "double"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isStarred": {
            "type": "boolean"
          },
          "mediaAttachmentType": {
            "type": "string",
            "nullable": true
          },
          "mediaAttachmentURL": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateToDoChecklistReqDto": {
        "type": "object",
        "properties": {
          "toDoChecklistId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "toDoTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}